- Move JFreeChart extensions that modifies/copy JFreeChart code directly to separate project, to comply with LGPL license without affect all IPED license scheme. - Update references to these moved classes. - Remove some unused references.
