--- iped-app/src/main/java/iped/viewers/timelinegraph/model/Minute.java
@@ -1,464 +1,0 @@
-package iped.viewers.timelinegraph.model;
-
-import java.io.Serializable;
-import java.util.Calendar;
-import java.util.Date;
-import java.util.Locale;
-import java.util.TimeZone;
-import org.jfree.chart.util.Args;
-import org.jfree.data.time.Day;
-import org.jfree.data.time.Hour;
-import org.jfree.data.time.RegularTimePeriod;
-
-/**
- * Represents a minute. This class is immutable, which is a requirement for all
- * {@link RegularTimePeriod} subclasses.
- * 
- * This class have to be implemented because of a bug in the original class with
- * the method equals, that do not considers the day to compare with the other
- * date, only the hour and the minute.
- */
-public class Minute extends RegularTimePeriod implements Serializable {
-
-    /** For serialization. */
-    private static final long serialVersionUID = 2144572840034842871L;
-
-    /** Useful constant for the first minute in a day. */
-    public static final int FIRST_MINUTE_IN_HOUR = 0;
-
-    /** Useful constant for the last minute in a day. */
-    public static final int LAST_MINUTE_IN_HOUR = 59;
-
-    /** The day. */
-    private Day day;
-
-    /** The hour in which the minute falls. */
-    private byte hour;
-
-    /** The minute. */
-    private byte minute;
-
-    /** The first millisecond. */
-    private long firstMillisecond;
-
-    /** The last millisecond. */
-    private long lastMillisecond;
-
-    /**
-     * Constructs a new Minute, based on the system date/time. The time zone and
-     * locale are determined by the calendar returned by
-     * {@link RegularTimePeriod#getCalendarInstance()}.
-     */
-    public Minute() {
-        this(new Date());
-    }
-
-    /**
-     * Constructs a new Minute. The time zone and locale are determined by the
-     * calendar returned by {@link RegularTimePeriod#getCalendarInstance()}.
-     *
-     * @param minute
-     *            the minute (0 to 59).
-     * @param hour
-     *            the hour ({@code null} not permitted).
-     */
-    public Minute(int minute, Hour hour) {
-        Args.nullNotPermitted(hour, "hour");
-        this.minute = (byte) minute;
-        this.hour = (byte) hour.getHour();
-        this.day = hour.getDay();
-        peg(getCalendarInstance());
-    }
-
-    /**
-     * Constructs a new instance, based on the supplied date/time. The time zone and
-     * locale are determined by the calendar returned by
-     * {@link RegularTimePeriod#getCalendarInstance()}.
-     *
-     * @param time
-     *            the time ({@code null} not permitted).
-     *
-     * @see #Minute(Date, TimeZone, Locale)
-     */
-    public Minute(Date time) {
-        // defer argument checking
-        this(time, getCalendarInstance());
-    }
-
-    /**
-     * Constructs a new Minute, based on the supplied date/time and timezone.
-     *
-     * @param time
-     *            the time ({@code null} not permitted).
-     * @param zone
-     *            the time zone ({@code null} not permitted).
-     * @param locale
-     *            the locale ({@code null} not permitted).
-     */
-    public Minute(Date time, TimeZone zone, Locale locale) {
-        Args.nullNotPermitted(time, "time");
-        Args.nullNotPermitted(zone, "zone");
-        Args.nullNotPermitted(locale, "locale");
-        Calendar calendar = Calendar.getInstance(zone, locale);
-        calendar.setTime(time);
-        int min = calendar.get(Calendar.MINUTE);
-        this.minute = (byte) min;
-        this.hour = (byte) calendar.get(Calendar.HOUR_OF_DAY);
-        this.day = new Day(time, zone, locale);
-        peg(calendar);
-    }
-
-    /**
-     * Constructs a new instance, based on a particular date/time. The time zone and
-     * locale are determined by the {@code calendar} parameter.
-     *
-     * @param time
-     *            the date/time ({@code null} not permitted).
-     * @param calendar
-     *            the calendar to use for calculations ({@code null} not permitted).
-     */
-    public Minute(Date time, Calendar calendar) {
-        Args.nullNotPermitted(time, "time");
-        Args.nullNotPermitted(calendar, "calendar");
-        calendar.setTime(time);
-        int min = calendar.get(Calendar.MINUTE);
-        this.minute = (byte) min;
-        this.hour = (byte) calendar.get(Calendar.HOUR_OF_DAY);
-        this.day = new Day(time, calendar);
-        peg(calendar);
-    }
-
-    /**
-     * Creates a new minute. The time zone and locale are determined by the calendar
-     * returned by {@link RegularTimePeriod#getCalendarInstance()}.
-     *
-     * @param minute
-     *            the minute (0-59).
-     * @param hour
-     *            the hour (0-23).
-     * @param day
-     *            the day (1-31).
-     * @param month
-     *            the month (1-12).
-     * @param year
-     *            the year (1900-9999).
-     */
-    public Minute(int minute, int hour, int day, int month, int year) {
-        this(minute, new Hour(hour, new Day(day, month, year)));
-    }
-
-    /**
-     * Returns the day.
-     *
-     * @return The day.
-     */
-    public Day getDay() {
-        return this.day;
-    }
-
-    /**
-     * Returns the hour.
-     *
-     * @return The hour (never {@code null}).
-     */
-    public Hour getHour() {
-        return new Hour(this.hour, this.day);
-    }
-
-    /**
-     * Returns the hour.
-     *
-     * @return The hour.
-     */
-    public int getHourValue() {
-        return this.hour;
-    }
-
-    /**
-     * Returns the minute.
-     *
-     * @return The minute.
-     */
-    public int getMinute() {
-        return this.minute;
-    }
-
-    /**
-     * Returns the first millisecond of the minute. This will be determined relative
-     * to the time zone specified in the constructor, or in the calendar instance
-     * passed in the most recent call to the {@link #peg(Calendar)} method.
-     *
-     * @return The first millisecond of the minute.
-     *
-     * @see #getLastMillisecond()
-     */
-    @Override
-    public long getFirstMillisecond() {
-        return this.firstMillisecond;
-    }
-
-    /**
-     * Returns the last millisecond of the minute. This will be determined relative
-     * to the time zone specified in the constructor, or in the calendar instance
-     * passed in the most recent call to the {@link #peg(Calendar)} method.
-     *
-     * @return The last millisecond of the minute.
-     *
-     * @see #getFirstMillisecond()
-     */
-    @Override
-    public long getLastMillisecond() {
-        return this.lastMillisecond;
-    }
-
-    /**
-     * Recalculates the start date/time and end date/time for this time period
-     * relative to the supplied calendar (which incorporates a time zone).
-     *
-     * @param calendar
-     *            the calendar ({@code null} not permitted).
-     */
-    @Override
-    public void peg(Calendar calendar) {
-        this.firstMillisecond = getFirstMillisecond(calendar);
-        this.lastMillisecond = getLastMillisecond(calendar);
-    }
-
-    /**
-     * Returns the minute preceding this one. No matter what time zone and locale
-     * this instance was created with, the returned instance will use the default
-     * calendar for time calculations, obtained with
-     * {@link RegularTimePeriod#getCalendarInstance()}.
-     *
-     * @return The minute preceding this one.
-     */
-    @Override
-    public RegularTimePeriod previous() {
-        Minute result;
-        if (this.minute != FIRST_MINUTE_IN_HOUR) {
-            result = new Minute(this.minute - 1, getHour());
-        } else {
-            Hour h = (Hour) getHour().previous();
-            if (h != null) {
-                result = new Minute(LAST_MINUTE_IN_HOUR, h);
-            } else {
-                result = null;
-            }
-        }
-        return result;
-    }
-
-    /**
-     * Returns the minute following this one. No matter what time zone and locale
-     * this instance was created with, the returned instance will use the default
-     * calendar for time calculations, obtained with
-     * {@link RegularTimePeriod#getCalendarInstance()}.
-     *
-     * @return The minute following this one.
-     */
-    @Override
-    public RegularTimePeriod next() {
-        Minute result;
-        if (this.minute != LAST_MINUTE_IN_HOUR) {
-            result = new Minute(this.minute + 1, getHour());
-        } else { // we are at the last minute in the hour...
-            Hour nextHour = (Hour) getHour().next();
-            if (nextHour != null) {
-                result = new Minute(FIRST_MINUTE_IN_HOUR, nextHour);
-            } else {
-                result = null;
-            }
-        }
-        return result;
-    }
-
-    /**
-     * Returns a serial index number for the minute.
-     *
-     * @return The serial index number.
-     */
-    @Override
-    public long getSerialIndex() {
-        long hourIndex = this.day.getSerialIndex() * 24L + this.hour;
-        return hourIndex * 60L + this.minute;
-    }
-
-    /**
-     * Returns the first millisecond of the minute.
-     *
-     * @param calendar
-     *            the calendar which defines the timezone ({@code null} not
-     *            permitted).
-     *
-     * @return The first millisecond.
-     *
-     * @throws NullPointerException
-     *             if {@code calendar} is {@code null}.
-     */
-    @Override
-    public long getFirstMillisecond(Calendar calendar) {
-        int year = this.day.getYear();
-        int month = this.day.getMonth() - 1;
-        int d = this.day.getDayOfMonth();
-
-        calendar.clear();
-        calendar.set(year, month, d, this.hour, this.minute, 0);
-        calendar.set(Calendar.MILLISECOND, 0);
-
-        return calendar.getTimeInMillis();
-    }
-
-    /**
-     * Returns the last millisecond of the minute.
-     *
-     * @param calendar
-     *            the calendar / timezone ({@code null} not permitted).
-     *
-     * @return The last millisecond.
-     *
-     * @throws NullPointerException
-     *             if {@code calendar} is {@code null}.
-     */
-    @Override
-    public long getLastMillisecond(Calendar calendar) {
-        int year = this.day.getYear();
-        int month = this.day.getMonth() - 1;
-        int d = this.day.getDayOfMonth();
-
-        calendar.clear();
-        calendar.set(year, month, d, this.hour, this.minute, 59);
-        calendar.set(Calendar.MILLISECOND, 999);
-
-        return calendar.getTimeInMillis();
-    }
-
-    /**
-     * Tests the equality of this object against an arbitrary Object.
-     * <P>
-     * This method will return true ONLY if the object is a Minute object
-     * representing the same minute as this instance.
-     *
-     * @param obj
-     *            the object to compare ({@code null} permitted).
-     *
-     * @return {@code true} if the minute and hour value of this and the object are
-     *         the same.
-     */
-    @Override
-    public boolean equals(Object obj) {
-        if (obj == this) {
-            return true;
-        }
-        if (!(obj instanceof Minute)) {
-            return false;
-        }
-        Minute that = (Minute) obj;
-        if (this.minute != that.minute) {
-            return false;
-        }
-        if (this.hour != that.hour) {
-            return false;
-        }
-        if (this.hour != that.hour) {
-            return false;
-        }
-        if (!this.day.equals(that.day)) {
-            return false;
-        }
-        return true;
-    }
-
-    /**
-     * Returns a hash code for this object instance. The approach described by
-     * Joshua Bloch in "Effective Java" has been used here:
-     * <p>
-     * {@code http://developer.java.sun.com/developer/Books/effectivejava
-     * /Chapter3.pdf}
-     *
-     * @return A hash code.
-     */
-    @Override
-    public int hashCode() {
-        int result = 17;
-        result = 37 * result + this.minute;
-        result = 37 * result + this.hour;
-        result = 37 * result + this.day.hashCode();
-        return result;
-    }
-
-    /**
-     * Returns an integer indicating the order of this Minute object relative to the
-     * specified object:
-     *
-     * negative == before, zero == same, positive == after.
-     *
-     * @param o1
-     *            object to compare.
-     *
-     * @return negative == before, zero == same, positive == after.
-     */
-    @Override
-    public int compareTo(Object o1) {
-        int result;
-
-        // CASE 1 : Comparing to another Minute object
-        // -------------------------------------------
-        if (o1 instanceof Minute) {
-            Minute m = (Minute) o1;
-            result = getHour().compareTo(m.getHour());
-            if (result == 0) {
-                result = this.minute - m.getMinute();
-            }
-        }
-
-        // CASE 2 : Comparing to another TimePeriod object
-        // -----------------------------------------------
-        else if (o1 instanceof RegularTimePeriod) {
-            // more difficult case - evaluate later...
-            result = 0;
-        }
-
-        // CASE 3 : Comparing to a non-TimePeriod object
-        // ---------------------------------------------
-        else {
-            // consider time periods to be ordered after general objects
-            result = 1;
-        }
-
-        return result;
-    }
-
-    /**
-     * Creates a Minute instance by parsing a string. The string is assumed to be in
-     * the format "YYYY-MM-DD HH:MM", perhaps with leading or trailing whitespace.
-     *
-     * @param s
-     *            the minute string to parse.
-     *
-     * @return {@code null}, if the string is not parseable, the minute otherwise.
-     */
-    public static Minute parseMinute(String s) {
-        Minute result = null;
-        s = s.trim();
-
-        String daystr = s.substring(0, Math.min(10, s.length()));
-        Day day = Day.parseDay(daystr);
-        if (day != null) {
-            String hmstr = s.substring(Math.min(daystr.length() + 1, s.length()), s.length());
-            hmstr = hmstr.trim();
-
-            String hourstr = hmstr.substring(0, Math.min(2, hmstr.length()));
-            int hour = Integer.parseInt(hourstr);
-
-            if ((hour >= 0) && (hour <= 23)) {
-                String minstr = hmstr.substring(Math.min(hourstr.length() + 1, hmstr.length()), hmstr.length());
-                int minute = Integer.parseInt(minstr);
-                if ((minute >= 0) && (minute <= 59)) {
-                    result = new Minute(minute, new Hour(hour, day));
-                }
-            }
-        }
-        return result;
-    }
-
-}
